// src/server/modules/reports/services/utils/different/interfaces/addDefaultMetricsToSku.interface.ts
export interface IDefaultSkuMetricsField {
  qty: number;
  tax: number;
  fines: number;
  netProfit: number;
  profitMargin: number;
  retailAmount: number;
  returnAmount: number;
  storageCost: number;
  deliveryCost: number;
  taxableAmount: number;
  acceptance: number;
  insuranceFee: number;
  otherExpenses: number;
  sellerPayoutAmount: number;
  deductionOrPayment: number;
  additionalInsuranceFee: number;
}